Django/requirements.txt
django-recaptcha
djangorestframework
djangorestframework-simplejwt
rest_framework_simplejwt
django-ckeditor
drf-yasg